flask-multi-db-monorepo/apps/azure-sql-app/requirements.txt
Flask>=2.2.0
mssql-python>=1.0.0
Flask-Cors>=3.0.10
azure-identity>=1.12.0
python-dotenv>=1.0.0
requests>=2.28.0